# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClipFlow API",
    "description": "A FastAPI backend service for video editing that analyzes audio tracks to detect speech segments, silence, and optionally filler words.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/upload": {
      "post": {
        "tags": ["Media"],
        "summary": "Upload Media",
        "description": "Upload a media file to a project.",
        "operationId": "upload_media_api_v1_projects__project_id__upload_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MediaFileResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project Not Found"
          },
          "413": {
            "description": "File Too Large"
          },
          "415": {
            "description": "Unsupported Media Type"
          }
        }
      }
    },
    "/api/v1/upload/presigned": {
      "post": {
        "tags": ["Media"],
        "summary": "Get Presigned Upload URL",
        "description": "Get a presigned URL for direct S3 upload.",
        "operationId": "get_presigned_upload_url_api_v1_upload_presigned_post",
        "security": [{ "HTTPBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresignedUploadRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PresignedUploadResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not Implemented (S3 not configured)"
          }
        }
      }
    },
    "/api/v1/upload/confirm/{file_id}": {
      "post": {
        "tags": ["Media"],
        "summary": "Confirm Upload",
        "description": "Confirm that a presigned URL upload has completed.",
        "operationId": "confirm_upload_api_v1_upload_confirm__file_id__post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadConfirmResponse" }
              }
            }
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/v1/media/{media_id}/waveform": {
      "get": {
        "tags": ["Media"],
        "summary": "Get Waveform",
        "description": "Get waveform data for a media file.",
        "operationId": "get_waveform_api_v1_media__media_id__waveform_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WaveformResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MediaFileResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "project_id": { "type": "string", "format": "uuid" },
          "original_filename": { "type": "string" },
          "stored_filename": { "type": "string" },
          "file_size": { "type": "integer" },
          "mime_type": { "type": "string" },
          "duration_seconds": { "type": "number", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "PresignedUploadRequest": {
        "type": "object",
        "required": ["filename", "content_type", "file_size"],
        "properties": {
          "filename": { "type": "string", "minLength": 1, "maxLength": 255 },
          "content_type": { "type": "string" },
          "file_size": { "type": "integer", "minimum": 1 }
        }
      },
      "PresignedUploadResponse": {
        "type": "object",
        "properties": {
          "upload_url": { "type": "string" },
          "file_id": { "type": "string", "format": "uuid" }
        }
      },
      "UploadConfirmResponse": {
        "type": "object",
        "properties": {
          "media_file": { "$ref": "#/components/schemas/MediaFileResponse" }
        }
      },
      "WaveformResponse": {
        "type": "object",
        "properties": {
          "peaks": {
            "type": "array",
            "items": { "type": "number" }
          },
          "duration": { "type": "number" },
          "sample_rate": { "type": "integer" }
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}